+++ TileEngineSfmlCs/TileEngineSfmlCs/TileEngine/TimeManagement/TimerTimeProvider.cs
@@ -7,8 +7,8 @@ namespace TileEngineSfmlCs.TileEngine.TimeManagement
     {
         private readonly Timer _timer;
         public event Action NextFrameEvent;
-        public float DeltaTime { get; private set; }
-        public float TotalTime { get; private set; }
+        public double DeltaTime { get; private set; }
+        public double TotalTime { get; private set; }
 
         private long _prevOsTicks;
